Fixed Enemy Collision Detection
