+++ Core/Entities/Characters/Character.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using ECS2022_23.Core.Animations;
+using ECS2022_23.Core.World;
 using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
@@ -13,6 +14,7 @@ public abstract class Character : Entity
     protected int SpriteWidth;
     public int AimDirection;
     public bool IsAlive = true;
+    public Level _level;
     public bool IsAttacking { get; set; }
     
     protected Character(Vector2 spawn, Texture2D texture) : base(spawn, texture)
@@ -33,6 +35,73 @@ public abstract class Character : Entity
     {
         
     }
+    
+    public bool Collides(Vector2 velocity)
+    {
+        var newPoint = (Position + velocity).ToPoint();
+        var rect = new Rectangle(newPoint, new Point(Texture.Width, Texture.Height));
+
+        //TODO clean up
+        
+        var armHitBoxLeft =
+            new Rectangle(newPoint.X + 4, newPoint.Y + Texture.Height / 2 + 2, 1, Texture.Height / 2 - 2);
+        var armHitBoxRight = new Rectangle(newPoint.X + Texture.Width - 5, newPoint.Y + Texture.Height / 2 + 2, 1,
+            Texture.Height / 2 - 2);
+
+        var feet = new Point(rect.Center.X, rect.Bottom);
+
+        if (velocity == Vector2.Zero)
+        {
+            return true;
+        }
+
+        var feetOnGround = false;
+
+        foreach (var rectangle in _level.GroundLayer)
+        {
+            if (rectangle.Contains(feet))
+            {
+                feetOnGround = true;
+            }
+        }
+
+        if (!feetOnGround) return false;
+
+        foreach (var rectangle in _level.GroundLayer)
+        {
+            if (velocity.Y == 0 && velocity.X > 0)
+            {
+                if (rectangle.Intersects(armHitBoxRight))
+                {
+                    return true;
+                }
+            }
+
+            if (velocity.Y == 0 && velocity.X < 0)
+            {
+                if (rectangle.Intersects(armHitBoxLeft))
+                {
+                    return true;
+                }
+            }
+
+            if ((velocity.X != 0 || !(velocity.Y > 0)) && (velocity.X != 0 || !(velocity.Y < 0))) continue;
+            
+            if (rectangle.Intersects(armHitBoxLeft) && rectangle.Intersects(armHitBoxRight))
+            {
+                return true;
+            }
+            
+        }
+        
+        return false;
+    }
+    
+    public void SetLevel(Level level)
+    {
+        this._level = level;
+    }
+    
     
     public override void Draw(SpriteBatch spriteBatch)
     {
